pyproject.toml
[project]
name = "bi-agent"
version = "0.1.0"
description = "数据分析智能代理 - 全自动/半自动数据分析工具"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "matplotlib>=3.7.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "httpx>=0.24.0",
    "socksio>=1.0.0",
    "typing-extensions>=4.5.0",
]

[project.scripts]
bi-agent = "bi_agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

